+++ test/transforms/test_line_digraph.py
@@ -1,0 +1,33 @@
+import torch
+
+from torch_geometric.data import Data
+from torch_geometric.transforms import LineDiGraph
+
+
+def test_line_digraph():
+    transform = LineDiGraph()
+    assert str(transform) == 'LineDiGraph()'
+
+    edge_index = torch.tensor([
+        [0, 1, 2, 2, 3],
+        [1, 2, 0, 3, 0],
+    ])
+    data = Data(edge_index=edge_index, num_nodes=4)
+    data = transform(data)
+    assert data.edge_index.tolist() == [[0, 1, 1, 2, 3, 4], [1, 2, 3, 0, 4, 0]]
+    assert data.num_nodes == data.edge_index.max().item() + 1
+
+    edge_index = torch.tensor([[0, 0, 0, 1, 2, 2, 3, 3, 4],
+                               [1, 2, 3, 4, 0, 3, 0, 4, 1]])
+    edge_attr = torch.ones(edge_index.size(1))
+    data = Data(edge_index=edge_index, edge_attr=edge_attr, num_nodes=5)
+    data = transform(data)
+    assert data.edge_index.max().item() + 1 == data.x.size(0)
+    assert data.edge_index.tolist() == [[
+        0, 1, 1, 2, 2, 3, 4, 4, 4, 5, 5, 6, 6, 6, 7, 8
+    ], [3, 4, 5, 6, 7, 8, 0, 1, 2, 6, 7, 0, 1, 2, 8, 3]]
+    assert data.num_nodes <= data.edge_index.max().item() + 1
+
+
+if __name__ == "__main__":
+    test_line_digraph()